alpinejs

alpine_array	R Documentation

Create a JavaScript empty array for Alpine.js state

Description

Helper that creates an empty JavaScript array ⁠[]⁠ without writing raw JavaScript. Use as an initial value in alpine_data() when you want to start with an empty list.

Usage

alpine_array()

Details

This function is equivalent to htmlwidgets::JS("[]") but signals intent more clearly and reads more naturally in R code.

Value

An htmlwidgets::JS object containing an empty JSON array. Suitable for use as an initial value in alpine_data().

See Also

Other JavaScript Helpers: alpine_object()

Examples

htmltools::tags$div(
  htmltools::tags$button("Add Todo") |> alpine_on("click", "todos.push({id: Date.now(), text: newTodo, done: false}); newTodo = ''"),
  htmltools::tags$input(type = "text", placeholder = "New todo") |> alpine_model("newTodo")
) |>
  alpine_data(
    todos = alpine_array(),
    newTodo = ""
  )