alpinejs

alpine_text	R Documentation

Set text content from a JavaScript expression

Description

Creates an x-text attribute that sets an element's text content to the result of a JavaScript expression. Alpine.js converts the expression result to a string and updates the text whenever state changes.

Usage

alpine_text(tag, expression, .cssSelector = NULL)

Arguments

tag
An htmltools tag object.

expression
Character string containing a JavaScript expression that evaluates to a value to display as text. Can reference state variables from alpine_data() and use computed properties or method calls.

.cssSelector
Optional CSS selector for targeting a nested element.

Details

Sets text content (safe from HTML injection). Expression results are converted to strings.

Value

The modified tag with x-text attribute appended.

See Also

Other Core Directives: alpine_bind(), alpine_cloak(), alpine_data(), alpine_data_js(), alpine_effect(), alpine_for(), alpine_html(), alpine_id(), alpine_if(), alpine_ignore(), alpine_init(), alpine_modelable(), alpine_on(), alpine_ref(), alpine_show(), alpine_teleport(), alpine_transition(), alpine_when()

Examples

htmltools::tags$div() |>
  alpine_text("itemCount")